LayoutTests/fast/dynamic/floating-to-non-floating.html

<!DOCTYPE html>
<html>
<head>
<link rel="stylesheet" type="text/css" href="resources/style-bug91665.css">
<script src="../../resources/check-layout.js"></script>
<script src="resources/helper-bug91665.js"></script>
</head>
<body onload="runTest('.non-floating');">
<h4>Testcase for<a href="https://bugs.webkit.org/show_bug.cgi?id=91665"> Bug #91665.</a><br>This testcase checks
if an element is properly placed with respect to its previous sibling when float:left changes to float:none for
various combinations of display property toggeled between inline-block and block for both testElement and its previous sibling.
</h4>
<p><b>Case 1</b>: &lt;Block Element&gt;&lt;Floating Block Element&gt; <b>To </b> &lt;Block Element&gt;&lt;Block Element&gt;</p>
<div class="wrapper">
  <div class="box"></div><div class="testElement" data-offset-y="20"></div>
</div>
<p><b>Case 2</b>: &lt;Block Element&gt;&lt;Floating Block Element&gt; <b>To </b> &lt;Block Element&gt;&lt;Inline Element&gt;</p>
<div class="wrapper">
  <div class="box"></div><div class="testElement" makeInline=true data-offset-y="20"></div>
</div>
<p><b>Case 3</b>: &lt;Block Element&gt;&lt;Floating Block Element&gt; <b>To </b> &lt;Inline Element&gt;&lt;Block Element&gt;</p>
<div class="wrapper">
  <div class="box" makeInline=true></div><div class="testElement" data-offset-y="23"></div>
</div>
<p><b>Case 4</b>: &lt;Block Element&gt;&lt;Floating Block Element&gt; <b>To </b> &lt;Inline Element&gt;&lt;Inline Element&gt;</p>
<div class="wrapper">
  <div class="box" makeInline=true></div><div class="testElement" makeInline=true data-offset-y="0"></div>
</div>
<p><b>Case 5</b>: &lt;Block Element&gt;&lt;Floating Inline Element&gt; <b>To </b> &lt;Block Element&gt;&lt;Block Element&gt;</p>
<div class="wrapper">
  <div class="box"></div><div class="testElement inline" data-offset-y="20"></div>
</div>
<p><b>Case 6</b>: &lt;Block Element&gt;&lt;Floating Block Element&gt; <b>To </b> &lt;Inline Element&gt;&lt;Block Element&gt;</p>
<div class="wrapper">
  <div class="box" makeInline=true></div><div class="testElement inline" makeBlock=true data-offset-y="23"></div>
</div>
<p><b>Case 7</b>: &lt;Block Element&gt;&lt;Floating Inline Element&gt; <b>To </b> &lt;Inline Element&gt;&lt;Inline Element&gt;</p>
<div class="wrapper">
  <div class="box" makeInline=true></div><div class="testElement inline" data-offset-y="0"></div>
</div>
<p><b>Case 8</b>: &lt;Inline Element&gt;&lt;Floating Block Element&gt; <b>To </b> &lt;Block Element&gt;&lt;Inline Element&gt;</p>
<div class="wrapper">
  <div class="box inline" makeBlock=true></div><div class="testElement" makeInline=true data-offset-y="20"></div>
</div>
<p><b>Case 9</b>: &lt;Inline Element&gt;&lt;Floating Block Element&gt; <b>To </b> &lt;Inline Element&gt;&lt;Block Element&gt;</p>
<div class="wrapper">
  <div class="box inline"></div><div class="testElement" data-offset-y="23"></div>
</div>
<p><b>Case 10</b>: &lt;Inline Element&gt;&lt;Floating Block Element&gt; <b>To </b> &lt;Inline Element&gt;&lt;Inline Element&gt;</p>
<div class="wrapper">
  <div class="box inline"></div><div class="testElement" makeInline=true data-offset-y="0"></div>
</div>
<p><b>Case 11</b>: &lt;Inline Element&gt;&lt;Floating Block Element&gt; <b>To </b> &lt;Block Element&gt;&lt;Inline Element&gt;</p>
<div class="wrapper">
  <div class="box inline" makeBlock=true></div><div class="testElement inline" data-offset-y="20"></div>
</div>
<p><b>Case 12</b>: &lt;Inline Element&gt;&lt;Floating Inline Element&gt; <b>To </b> &lt;Inline Element&gt;&lt;Block Element&gt;</p>
<div class="wrapper">
  <div class="box inline"></div><div class="testElement inline" makeBlock=true data-offset-y="23"></div>
</div>
<p><b>Case 13</b>: &lt;Inline Element&gt;&lt;Floating Inline Element&gt; <b>To </b> &lt;Inline Element&gt;&lt;Inline Element&gt;</p>
<div class="wrapper">
  <div class="box inline"></div><div class="testElement inline" data-offset-y="0"></div>
</div>
<p><b>Case 14</b>: &lt;Block Element&gt;&lt;Floating Table Element&gt; <b>To </b> &lt;Block Element&gt;&lt;Table Element&gt;</p>
<div class="wrapper">
  <div class="box"></div><table class="testElement" data-offset-y="20"></table>
</div>
<p><b>Case 15</b>: &lt;Block Element&gt;&lt;Floating Table Element&gt; <b>To </b> &lt;Block Element&gt;&lt;Inline-Table Element&gt;</p>
<div class="wrapper">
  <div class="box"></div><table class="testElement" makeInline=true data-offset-y="20"></table>
</div>
<p><b>Case 16</b>: &lt;Block Element&gt;&lt;Floating Table Element&gt; <b>To </b> &lt;Inline Element&gt;&lt;Table Element&gt;</p>
<div class="wrapper">
  <div class="box" makeInline=true></div><table class="testElement" data-offset-y="23"></table>
</div>
<p><b>Case 17</b>: &lt;Block Element&gt;&lt;Floating Table Element&gt; <b>To </b> &lt;Inline Element&gt;&lt;Inline-Table Element&gt;</p>
<div class="wrapper">
  <div class="box" makeInline=true></div><table class="testElement" makeInline=true data-offset-y="0"></table>
</div>
<p><b>Case 18</b>: &lt;Block Element&gt;&lt;Floating Inline-Table Element&gt; <b>To </b> &lt;Block Element&gt;&lt;Inline-Table Element&gt;</p>
<div class="wrapper">
  <div class="box"></div><table class="testElement inline" data-offset-y="20"></table>
</div>
<p><b>Case 19</b>: &lt;Block Element&gt;&lt;Floating Inline Element&gt; <b>To </b> &lt;Block Element&gt;&lt;Table Element&gt;</p>
<div class="wrapper">
  <div class="box" makeInline=true></div><table class="testElement inline" makeBlock=true data-offset-y="23"></table>
</div>
<p><b>Case 20</b>: &lt;Block Element&gt;&lt;Floating Inline-Table Element&gt; <b>To </b> &lt;Inline Element&gt;&lt;Inline-Table Element&gt;</p>
<div class="wrapper">
  <div class="box" makeInline=true></div><table class="testElement inline" data-offset-y="0"></table>
</div>
<p><b>Case 21</b>: &lt;Inline Element&gt;&lt;Floating Table Element&gt; <b>To </b> &lt;Inline Element&gt;&lt;Inline-Table Element&gt;</p>
<div class="wrapper">
  <div class="box inline" makeBlock=true></div><table class="testElement" makeInline=true data-offset-y="20"></table>
</div>
<p><b>Case 22</b>: &lt;Inline Element&gt;&lt;Floating Table Element&gt; <b>To </b> &lt;Inline Element&gt;&lt;Table Element&gt;</p>
<div class="wrapper">
  <div class="box inline"></div><table class="testElement" data-offset-y="23"></table>
</div>
<p><b>Case 23</b>: &lt;Inline Element&gt;&lt;Floating Block Element&gt; <b>To </b> &lt;Inline Element&gt;&lt;Inline-Table Element&gt;</p>
<div class="wrapper">
  <div class="box inline"></div><table class="testElement" makeInline=true data-offset-y="0"></table>
</div>
<p><b>Case 24</b>: &lt;Inline Element&gt;&lt;Floating Inline-Table Element&gt; <b>To </b> &lt;Block Element&gt;&lt;Inline-Table Element&gt;</p>
<div class="wrapper">
  <div class="box inline" makeBlock=true></div><table class="testElement inline" data-offset-y="20"></table>
</div>
<p><b>Case 25</b>: &lt;Inline Element&gt;&lt;Floating Inline-Table Element&gt; <b>To </b> &lt;Inline Element&gt;&lt;Table Element&gt;</p>
<div class="wrapper">
  <div class="box inline"></div><table class="testElement inline" makeBlock=true data-offset-y="23"></table>
</div>
<p><b>Case 26</b>: &lt;Inline Element&gt;&lt;Floating Inline-Table Element&gt; <b>To </b> &lt;Inline Element&gt;&lt;Inline-Table Element&gt;</p>
<div class="wrapper">
  <div class="box inline"></div><table class="testElement inline" data-offset-y="0"></table>
</div>
</body>
</html>
